PBF: convert u8 to float for left joysticks in Flying Trial Farmer

+++ SerialPrograms/Source/PokemonSV/Programs/Farming/PokemonSV_FlyingTrialFarmer.h
@@ -51,7 +51,7 @@ private:
     EventNotificationsOption NOTIFICATIONS;
 
     bool run_rewards(SingleSwitchProgramEnvironment& env, ProControllerContext& context);
-    uint8_t get_final_y_axis(int8_t delta_y);
+    double get_final_y_axis(double delta_y);
 };
